Trading-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0
python-multipart
kiteconnect
pydantic[email]
pydantic-settings
python-dotenv
requests
